+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog for PSDI Data Conversion
 
+## v0.4.0
+
+### New and Changed Functionality
+
+- The database method `get_converter_info` in can now be called without a `name` argument, and will return a list of info on all converters
+- The converter info provided by queries to the database methods now contains member variables `supported` and `registered` indicating the status of the converter:
+  - Both `False`: The converter is known to exist, but we currently provide no support for it
+  - `supported==True`, `registered==False`: This package has a plugin ready to support this converter, but it cannot currently be used due to e.g. a required binary being missing which must be provided by the user
+  - Both `True`: The converter is supported and ready to use
+
+### Documentation Changes
+
+- Fixed incorrect output type in documentation for database methods `get_in_format_args` and `get_out_format_args`
+- Improved the error output for the database method `get_converter_info` if the converter name is not recognised, providing a list of known names
+
 ## v0.3.23
 
 ### Bugfixes